+++ test/unit/braintree/webhook_notification_gateway_spec.js
@@ -833,5 +833,23 @@ describe('WebhookNotificationGateway', function () {
         done();
       });
     });
+
+    it('returns a parsable signature and payload for Local Payment Reversed', function (done) {
+      let notification = specHelper.defaultGateway.webhookTesting.sampleNotification(
+        WebhookNotification.Kind.LocalPaymentReversed,
+        'my_id'
+      );
+      let bt_signature = notification.bt_signature;
+      let bt_payload = notification.bt_payload;
+
+      specHelper.defaultGateway.webhookNotification.parse(bt_signature, bt_payload, function (err, webhookNotification) {
+        assert.equal(webhookNotification.kind, WebhookNotification.Kind.LocalPaymentReversed);
+
+        let localPaymentReversed = webhookNotification.subject.localPaymentReversed;
+
+        assert.equal('a-payment-id', localPaymentReversed.paymentId);
+        done();
+      });
+    });
   });
 });